move tests to standard Tests namespace
