+++ composer.json
@@ -29,7 +29,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "GitWrapper\\Test\\": "tests"
+            "GitWrapper\\Tests\\": "tests"
         }
     },
     "suggest": {
