+++ tests/Event/TestOutputListener.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GitWrapper\Test\Event;
+namespace GitWrapper\Tests\Event;
 
 use GitWrapper\Contract\Event\GitOutputListenerInterface;
 use GitWrapper\Event\GitOutputEvent;
