+++ tests/EventSubscriber/GitLoggerEventSubscriberTest.php
@@ -2,11 +2,13 @@
 
 declare(strict_types=1);
 
-namespace GitWrapper\Test;
+namespace GitWrapper\Tests\EventSubscriber;
 
 use GitWrapper\EventSubscriber\GitLoggerEventSubscriber;
 use GitWrapper\Exception\GitException;
 use GitWrapper\GitCommand;
+use GitWrapper\Tests\AbstractGitWrapperTestCase;
+use GitWrapper\Tests\Log\TestLogger;
 use Psr\Log\LogLevel;
 use Psr\Log\NullLogger;
 use Throwable;
