+++ tests/GitWrapperTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace GitWrapper\Test;
+namespace GitWrapper\Tests;
 
 use GitWrapper\Exception\GitException;
 use GitWrapper\GitCommand;
 use GitWrapper\GitWorkingCopy;
 use GitWrapper\GitWrapper;
-use GitWrapper\Test\Event\TestDispatcher;
+use GitWrapper\Tests\Event\TestDispatcher;
 
 final class GitWrapperTest extends AbstractGitWrapperTestCase
 {
